requirements.txt
# Django核心框架
Django==4.2.10

# 环境变量管理
python-dotenv==1.0.1

# 数据库相关
mysqlclient==2.2.0
pymysql==1.1.1

# Web服务器
gunicorn==21.2.0

# 静态文件服务
whitenoise==6.6.0

# HTTP请求库
requests==2.31.0

# 企业微信SDK
wechatpy==1.8.18

# 加密库 (wechatpy依赖)
cryptography==45.0.3

# Excel文件处理
openpyxl==3.1.2

# 用户代理解析
user-agents==2.2.0

# 任务队列
celery==5.3.4
django-celery-beat==2.5.0

# CORS支持
django-cors-headers==4.3.1

# Redis (Celery broker)
redis==5.0.1

# 时区处理
pytz==2023.3
